% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polylabel.R
\name{centroid}
\alias{centroid}
\title{Title}
\usage{
centroid(x, y = NULL, precision = 1, ...)
}
\arguments{
\item{x}{a vector of x coordinates or a matrix or data.frame
of x and y coordinates, a list of components x and y, or
a time series (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}} for details)}

\item{y}{a vector of y coordinates. Only needs to be provided if
\code{x} is vector.}

\item{precision}{the precision to use when computing the center}

\item{...}{arguments passed down to \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}}
}
\value{
A list with items \code{x} and \code{y}, representing the
centroid of the polygon(s).
}
\description{
Title
}
\examples{
x <- rbind(c(2, 2),
           c(2.5, 2),
           c(2.5, 2.5),
           c(2, 2.5),
           c(NA, NA),
           c(0, 0),
           c(0, 1),
           c(1, 1),
           c(1, 0))
centroid <- centroid(x, precision = 0.5)
plot(x, type = 'n')
polygon(x)
points(centroid)

}
\seealso{
\code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}, \code{\link[graphics:polygon]{graphics::polygon()}}
}
