% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polylabelr.R
\docType{package}
\name{polylabelr-package}
\alias{polylabelr}
\alias{polylabelr-package}
\title{polylabelr: Efficiently Locate the Center of a Polygon for Labeling}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/polylabelr}
  \item Report bugs at \url{https://github.com/jolars/polylabelr/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{mail@larssonjohan.com}

Other contributors:
\itemize{
  \item Mapbox (mapbox libraries, including polylabel) [copyright holder]
}

}
\keyword{internal}
