% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polylabelr.R
\docType{package}
\name{polylabelr-package}
\alias{polylabelr}
\alias{polylabelr-package}
\title{polylabelr: Find the Pole of Inaccessibility (Visual Center) of a Polygon}
\description{
A wrapper around the C++ library 
    'polylabel' from 'Mapbox', providing an efficient routine for
    finding the approximate pole of inaccessibility of a polygon, which
    usually serves as an excellent candidate for labeling of a polygon.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/polylabelr}
  \item Report bugs at \url{https://github.com/jolars/polylabelr/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{mail@larssonjohan.com}

Other contributors:
\itemize{
  \item Mapbox (variant, mapbox, and geometry libraries) [copyright holder]
}

}
\keyword{internal}
